// src/services/ebird/types/EbirdLocation.ts
export default interface EbirdLocation {
  countryCode: string;
  countryName: string;
  hierarchicalName: string;
  isHotspot: boolean;
  lat: number;
  latitude: number;
  lng: number;
  locID: string;
  locId: string;
  locName: string;
  longitude: number;
  name: string;
  subnational1Code: string;
  subnational1Name: string;
  subnational2Code: string;
  subnational2Name: string;
}