// src/services/ebird/types/EbirdObservation.ts
export default interface EbirdObservation {
  checklistId?: string;
  comName: string;
  countryCode?: string;
  countryName?: string;
  evidence?: string;
  exoticCategory?: string;
  firstName?: string;
  hasComments?: boolean;
  hasRichMedia?: boolean;
  howMany?: number;
  lastName?: string;
  lat: number;
  lng: number;
  locId: string;
  locName: string;
  locationPrivate: boolean;
  obsDt: string;
  obsReviewed: boolean;
  obsValid: boolean;
  presenceNoted?: boolean;
  sciName: string;
  speciesCode: string;
  subId: string;
  subnational1Code?: string;
  subnational1Name?: string;
  subnational2Code?: string;
  subnational2Name?: string;
  userDisplayName?: string;
}